archive/requirements.txt
# Core dependencies for PromptPlot v2.0
pydantic>=2.0.0
llama-index>=0.10.0
matplotlib>=3.5.0
opencv-python>=4.5.0
pyserial>=3.5
asyncio-mqtt>=0.11.0
aiofiles>=0.8.0
pillow>=9.0.0
numpy>=1.21.0

# Optional LLM providers
openai>=1.0.0
azure-identity>=1.12.0
ollama>=0.1.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.991